+++ pkg/alertmanager/multitenant.go
@@ -66,6 +66,7 @@ var (
 	errShardingLegacyStorage               = errors.New("deprecated -alertmanager.storage.* not supported with -alertmanager.sharding-enabled, use -alertmanager-storage.*")
 	errShardingUnsupportedStorage          = errors.New("the configured alertmanager storage backend is not supported when sharding is enabled")
 	errZoneAwarenessEnabledWithoutZoneInfo = errors.New("the configured alertmanager has zone awareness enabled but zone is not set")
+	errNotUploadingFallback                = errors.New("not uploading fallback configuration")
 )
 
 // MultitenantAlertmanagerConfig is the configuration for a multitenant Alertmanager.
@@ -1023,8 +1024,12 @@ func (am *MultitenantAlertmanager) serveRequest(w http.ResponseWriter, req *http
 	}
 
 	if am.fallbackConfig != "" {
-		userAM, err = am.alertmanagerFromFallbackConfig(userID)
-		if err != nil {
+		userAM, err = am.alertmanagerFromFallbackConfig(req.Context(), userID)
+		if errors.Is(err, errNotUploadingFallback) {
+			level.Warn(am.logger).Log("msg", "not initializing Alertmanager", "user", userID, "err", err)
+			http.Error(w, "Not initializing the Alertmanager", http.StatusNotAcceptable)
+			return
+		} else if err != nil {
 			level.Error(am.logger).Log("msg", "unable to initialize the Alertmanager with a fallback configuration", "user", userID, "err", err)
 			http.Error(w, "Failed to initialize the Alertmanager", http.StatusInternalServerError)
 			return
@@ -1038,10 +1043,32 @@ func (am *MultitenantAlertmanager) serveRequest(w http.ResponseWriter, req *http
 	http.Error(w, "the Alertmanager is not configured", http.StatusNotFound)
 }
 
-func (am *MultitenantAlertmanager) alertmanagerFromFallbackConfig(userID string) (*Alertmanager, error) {
+func (am *MultitenantAlertmanager) alertmanagerFromFallbackConfig(ctx context.Context, userID string) (*Alertmanager, error) {
+	// Make sure we never create fallback instances for a user not owned by this instance.
+	// This check is not strictly necessary as the configuration polling loop will deactivate
+	// any tenants which are not meant to be in an instance, but it is confusing and potentially
+	// wasteful to have them start-up when they are not needed.
+	if !am.isUserOwned(userID) {
+		return nil, errors.Wrap(errNotUploadingFallback, "user not owned by this instance")
+	}
+
+	// We should be careful never to replace an existing configuration with the fallback.
+	// There is a small window of time between the check and upload where a user could
+	// have uploaded a configuration, but this only applies to the first ever request made.
+	_, err := am.store.GetAlertConfig(ctx, userID)
+	if err == nil {
+		// If there is a configuration, then the polling cycle should pick it up.
+		return nil, errors.Wrap(errNotUploadingFallback, "user has a configuration")
+	}
+	if !errors.Is(err, alertspb.ErrNotFound) {
+		return nil, errors.Wrap(err, "failed to check for existing configuration")
+	}
+
+	level.Warn(am.logger).Log("msg", "no configuration exists for user; uploading fallback configuration", "user", userID)
+
 	// Upload an empty config so that the Alertmanager is no de-activated in the next poll
 	cfgDesc := alertspb.ToProto("", nil, userID)
-	err := am.store.SetAlertConfig(context.Background(), cfgDesc)
+	err = am.store.SetAlertConfig(ctx, cfgDesc)
 	if err != nil {
 		return nil, err
 	}