+++ pkg/alertmanager/multitenant_test.go
@@ -964,6 +964,84 @@ receivers:
 	require.Equal(t, http.StatusOK, resp.StatusCode)
 }
 
+// This test checks that the fallback configuration does not overwrite a configuration
+// written to storage before it is picked up by the instance.
+func TestMultitenantAlertmanager_ServeHTTPBeforeSyncFailsIfConfigExists(t *testing.T) {
+	ctx := context.Background()
+	amConfig := mockAlertmanagerConfig(t)
+
+	// Prevent polling configurations, we want to test the window of time
+	// between the configuration existing and it being incorporated.
+	amConfig.PollInterval = time.Hour
+
+	// Run this test using a real storage client.
+	store := prepareInMemoryAlertStore()
+
+	externalURL := flagext.URLValue{}
+	err := externalURL.Set("http://localhost:8080/alertmanager")
+	require.NoError(t, err)
+
+	fallbackCfg := `
+global:
+  smtp_smarthost: 'localhost:25'
+  smtp_from: 'youraddress@example.org'
+route:
+  receiver: example-email
+receivers:
+  - name: example-email
+    email_configs:
+    - to: 'youraddress@example.org'
+`
+	amConfig.ExternalURL = externalURL
+
+	// Create the Multitenant Alertmanager.
+	am, err := createMultitenantAlertmanager(amConfig, nil, nil, store, nil, nil, log.NewNopLogger(), nil)
+	require.NoError(t, err)
+	am.fallbackConfig = fallbackCfg
+
+	require.NoError(t, services.StartAndAwaitRunning(ctx, am))
+	defer services.StopAndAwaitTerminated(ctx, am) //nolint:errcheck
+
+	// Upload config for the user.
+	require.NoError(t, store.SetAlertConfig(ctx, alertspb.AlertConfigDesc{
+		User:      "user1",
+		RawConfig: simpleConfigOne,
+		Templates: []*alertspb.TemplateDesc{},
+	}))
+
+	// Request before the user configuration loaded by polling loop.
+	req := httptest.NewRequest("GET", externalURL.String()+"/api/v1/status", nil)
+	w := httptest.NewRecorder()
+	am.ServeHTTP(w, req.WithContext(user.InjectOrgID(req.Context(), "user1")))
+	resp := w.Result()
+	assert.Equal(t, http.StatusNotAcceptable, resp.StatusCode)
+
+	// Check the configuration has not been replaced.
+	readConfigDesc, err := store.GetAlertConfig(ctx, "user1")
+	require.NoError(t, err)
+	assert.Equal(t, simpleConfigOne, readConfigDesc.RawConfig)
+
+	// We expect the request to fail because the user has already uploaded a
+	// configuration and should not replace it.
+	assert.Len(t, am.alertmanagers, 0)
+
+	// Now force a poll to actually load the configuration.
+	err = am.loadAndSyncConfigs(ctx, reasonPeriodic)
+	require.NoError(t, err)
+
+	// Now it should exist. This is to sanity check the test is working as expected, and
+	// that the user configuration was written correctly such that it can be picked up.
+	require.Len(t, am.alertmanagers, 1)
+	_, exists := am.alertmanagers["user1"]
+	require.True(t, exists)
+
+	// Request should now succeed.
+	w = httptest.NewRecorder()
+	am.ServeHTTP(w, req.WithContext(user.InjectOrgID(req.Context(), "user1")))
+	resp = w.Result()
+	require.Equal(t, http.StatusOK, resp.StatusCode)
+}
+
 func TestMultitenantAlertmanager_InitialSyncWithSharding(t *testing.T) {
 	tc := []struct {
 		name          string